+++ sleap_io/io/__init__.py
@@ -1,1 +1,3 @@
 """This sub-package contains I/O-related modules such as specific format backends."""
+
+from . import video_reading as video
